requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
msal==1.24.1
requests==2.31.0
gunicorn==21.2.0
selenium==4.15.0